requirements.txt
django==2.1.1
django-oscar==1.6.5
# django-oscar==2.0 upgrade to later
mysqlclient==1.3.14
# mysql-connector-python==8.0.12
django-compressor==2.4